Started working on reworking the structure of slash commands
