+++ LostArkBot/Src/Bot/MenusOld/AbyssalDungeonMenu.cs
@@ -5,7 +5,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace LostArkBot.Src.Bot.Menus
+namespace LostArkBot.Src.Bot.MenusOld
 {
     internal class AbyssalDungeonMenu
     {
