+++ LostArkBot/Src/Bot/MenusOld/JoinCharacterMenu.cs
@@ -8,7 +8,7 @@ using System.Linq;
 using System.Text.Json;
 using System.Threading.Tasks;
 
-namespace LostArkBot.Src.Bot.Menus
+namespace LostArkBot.Src.Bot.MenusOld
 {
     internal class JoinCharacterMenu
     {
