+++ LostArkBot/Src/Bot/SlashCommandsOld/ProfileInitialization.cs
@@ -1,21 +1,21 @@
 ﻿using Discord;
 
-namespace LostArkBot.Src.Bot.SlashCommands
+namespace LostArkBot.Src.Bot.SlashCommandsOld
 {
-    internal class DeleteInitialization
+    internal class ProfileInitialization
     {
-        public static SlashCommandBuilder Delete()
+        public static SlashCommandBuilder Profile()
         {
-            SlashCommandBuilder deleteCommand = new SlashCommandBuilder()
-                                                     .WithName("delete")
-                                                     .WithDescription("Deletes the character")
+            SlashCommandBuilder profileCommand = new SlashCommandBuilder()
+                                                     .WithName("profile")
+                                                     .WithDescription("Shows the profile of the character")
                                                      .AddOption(new SlashCommandOptionBuilder()
                                                                 .WithName("character-name")
                                                                 .WithDescription("Name of the character")
                                                                 .WithRequired(true)
                                                                 .WithType(ApplicationCommandOptionType.String));
 
-            return deleteCommand;
+            return profileCommand;
         }
     }
 }
