+++ LostArkBot/Src/Bot/SlashCommandsOld/RegisterMetaInitialization.cs
@@ -1,21 +1,21 @@
 ﻿using Discord;
 
-namespace LostArkBot.Src.Bot.SlashCommands
+namespace LostArkBot.Src.Bot.SlashCommandsOld
 {
-    internal class UpdateMetaInitialization
+    internal class RegisterMetaInitialization
     {
-        public static SlashCommandBuilder UpdateMeta()
+        public static SlashCommandBuilder RegisterMeta()
         {
-            SlashCommandBuilder updateMetaCommand = new SlashCommandBuilder()
-                                                    .WithName("updatemeta")
-                                                    .WithDescription("Updates your currently logged in character with information from meta-game")
+            SlashCommandBuilder registerMetaCommand = new SlashCommandBuilder()
+                                                    .WithName("registermeta")
+                                                    .WithDescription("Registers your currently logged in character with information from meta-game")
                                                     .AddOption(new SlashCommandOptionBuilder()
                                                                .WithName("twitch-name")
                                                                .WithDescription("Name of your character, caps sensitive!")
                                                                .WithRequired(true)
                                                                .WithType(ApplicationCommandOptionType.String));
 
-            return updateMetaCommand;
+            return registerMetaCommand;
         }
     }
 }
