+++ LostArkBot/Src/Bot/SlashCommandsOld/ServerStatusInitialization.cs
@@ -1,6 +1,6 @@
 ﻿using Discord;
 
-namespace LostArkBot.Src.Bot.SlashCommands
+namespace LostArkBot.Src.Bot.SlashCommandsOld
 {
     internal class ServerStatusInitialization
     {
