+++ LostArkBot/Src/Bot/SlashCommandsOld/UpdateMetaInitialization.cs
@@ -1,21 +1,21 @@
 ﻿using Discord;
 
-namespace LostArkBot.Src.Bot.SlashCommands
+namespace LostArkBot.Src.Bot.SlashCommandsOld
 {
-    internal class RegisterMetaInitialization
+    internal class UpdateMetaInitialization
     {
-        public static SlashCommandBuilder RegisterMeta()
+        public static SlashCommandBuilder UpdateMeta()
         {
-            SlashCommandBuilder registerMetaCommand = new SlashCommandBuilder()
-                                                    .WithName("registermeta")
-                                                    .WithDescription("Registers your currently logged in character with information from meta-game")
+            SlashCommandBuilder updateMetaCommand = new SlashCommandBuilder()
+                                                    .WithName("updatemeta")
+                                                    .WithDescription("Updates your currently logged in character with information from meta-game")
                                                     .AddOption(new SlashCommandOptionBuilder()
                                                                .WithName("twitch-name")
                                                                .WithDescription("Name of your character, caps sensitive!")
                                                                .WithRequired(true)
                                                                .WithType(ApplicationCommandOptionType.String));
 
-            return registerMetaCommand;
+            return updateMetaCommand;
         }
     }
 }
